@@ -1,5 +1,9 @@
-name: leds
+name: leds_remasterized
 description: A new Flutter project.
+
+# The following line prevents the package from being accidentally published to
+# pub.dev using `pub publish`. This is preferred for private packages.
+publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 
 # The following defines the version and build number for your application.
 # A version number is three numbers separated by dots, like 1.2.43
@@ -14,26 +18,26 @@ description: A new Flutter project.
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  http: ^0.12.0+4
-  flutter_colorpicker: ^0.3.5
-  shared_preferences: ^0.5.12+4
-  easy_localization: ^2.3.3+1
-  provider: ^4.3.2+3
-  flutter_easyloading: ^2.2.2
+
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^0.1.2
+  cupertino_icons: ^1.0.2
+  flutter_easyloading: ^3.0.0
+  provider: ^5.0.0
+  easy_localization: ^3.0.0
+  shared_preferences: ^2.0.6
+  flutter_colorpicker: ^0.4.0
+  http: ^0.13.3
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
